feat : 병합 작업 (데이터 : 플레이어 스텟)

+++ Assets/Scripts/Character/CharacterData.cs
@@ -16,7 +16,7 @@ public class CharacterData : ScriptableObject
     public float attackDelay;
     public float attackSpeed;
     public float attackRange;
-    public float attackCount;
+    public int attackCount;
     public float attackAngle;
     public float knockbackForce;
     public float attackDuration;
